+++ __tests__/parsing/linux.test.tsx
@@ -32,5 +32,6 @@ rx bitrate: 780.0 MBit/s
     txRate: 866.7,
     phyMode: "",
     security: "",
+    active: false,
   });
 });